u32 instead of usize

+++ pallets/process-validation/src/weights.rs
@@ -2,7 +2,7 @@ use frame_support::{traits::Get, weights::Weight};
 use sp_std::marker::PhantomData;
 
 pub trait WeightInfo {
-    fn create_process(i: usize) -> Weight;
+    fn create_process(i: u32) -> Weight;
     fn disable_process() -> Weight;
 }
 
@@ -11,7 +11,7 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
     // Storage: ProcessValidation VersionModel (r:1 w:1)
     // Storage: ProcessValidation ProcessModel (r:1 w:1)
     /// The range of component `i` is `[1, 200]`.
-    fn create_process(i: usize) -> Weight {
+    fn create_process(i: u32) -> Weight {
         Weight::from_ref_time(25_000_000 as u64)
             // Standard Error: 156
             .saturating_add(Weight::from_ref_time(91_785 as u64).saturating_mul(i as u64))
@@ -28,7 +28,7 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 }
 
 impl WeightInfo for () {
-    fn create_process(_: usize) -> Weight {
+    fn create_process(_: u32) -> Weight {
         Weight::from_ref_time(0 as u64)
     }
     fn disable_process() -> Weight {
